+++ freecache/docs.go
@@ -1,0 +1,4 @@
+/*
+Package freecache contains an implementation of the `gokv.Store` interface for FreeCache.
+*/
+package freecache