+++ RELEASES.md
@@ -9,6 +9,7 @@ vNext
 -----
 
 - Added: Package `tablestorage` - A `gokv.Store` implementation for [Azure Table Storage](https://azure.microsoft.com/en-us/services/storage/tables/) (issue [#42](https://github.com/philippgille/gokv/issues/42))
+- Added: Package `freecache` - A `gokv.Store` implementation for [freecache](https://github.com/coocood/freecache) (issue [#44](https://github.com/philippgille/gokv/issues/44))
 
 v0.4.0 (2018-12-02)
 -------------------